add stopping

+++ test/test_stop.sh
@@ -2,11 +2,12 @@
 # shellcheck disable=SC1091,SC2086,SC2046
 set -e
 
-# stop any upgrade tests
-set -o allexport; source upgrade/default_variables.env; set +o allexport
-#docker-compose -f docker-compose.upgrade.yml down --volumes --remove-orphans
 
-DEFAULT_SUITE_CONFIG="-f docker-compose.yml"
+DEFAULT_SUITE_CONFIG=$1
+
+if [[ -z "${DEFAULT_SUITE_CONFIG// }" ]]; then
+    DEFAULT_SUITE_CONFIG="-f docker-compose.yml"
+fi
 
 ALL_SUITES_FILES="$(find suite-config/ -name "docker-compose.override.yml")"
 ALL_SUITES=""
@@ -14,9 +15,16 @@ for FILE in $ALL_SUITES_FILES; do
     ALL_SUITES="$ALL_SUITES -f $FILE"
 done
 
-
 set +e
+
+CONTAINERS=$(docker-compose $DEFAULT_SUITE_CONFIG $ALL_SUITES -f docker-compose-selenium-test.yml ps -q)
+
 # stop any suite things
-docker kill $(docker-compose $DEFAULT_SUITE_CONFIG $ALL_SUITES -f docker-compose-selenium-test.yml ps -q) > /dev/null
-docker-compose $DEFAULT_SUITE_CONFIG $ALL_SUITES -f docker-compose-selenium-test.yml down --volumes
+if [ -z "$CONTAINERS" ]; then
+    echo "No containers to terminate."
+else
+    docker kill $CONTAINERS > /dev/null
+    docker-compose $DEFAULT_SUITE_CONFIG $ALL_SUITES -f docker-compose-selenium-test.yml down --volumes
+fi
+
 set -e
